+++ packages/theme_tailor/lib/src/util/extension/field_declaration_extension.dart
@@ -1,7 +1,7 @@
 import 'package:analyzer/dart/ast/ast.dart';
 
 extension FieldDeclarationExtension on FieldDeclaration {
-  String get name => fields.variables.first.name.name;
+  String get name => fields.variables.first.name.toString();
 
   List<String> get annotations {
     return metadata.map((e) => e.toString()).toList(growable: false);