+++ packages/theme_tailor/lib/src/util/import_finder.dart
@@ -14,7 +14,7 @@ ElementMatcher isWithinLibrary(String name) {
 }
 
 class ImportFinder {
-  const ImportFinder({
+  ImportFinder({
     required this.lib,
     required this.whereElement,
     this.whereLibrary,
@@ -28,7 +28,7 @@ class ImportFinder {
   /// lib/src/dart/element/element.dart
   bool recursiveSearch() {
     final libraries = <LibraryElementWithVisibility>{};
-    for (final import in lib.imports) {
+    for (final import in lib.libraryImports) {
       final library = import.importedLibrary;
       if (library != null) {
         libraries.add(LibraryElementWithVisibility.root(
@@ -38,17 +38,29 @@ class ImportFinder {
       }
     }
 
-    return libraries.any(_isExportedRecursivaly);
+    return libraries.any((lib) => _isExportedRecursivaly(
+          lib,
+          {lib.lib.librarySource.fullName},
+        ));
   }
 
-  bool _isExportedRecursivaly(LibraryElementWithVisibility lib) {
+  bool _isExportedRecursivaly(
+    LibraryElementWithVisibility lib,
+    Set<String> visited,
+  ) {
     if (lib.isRelevant(whereLibrary)) {
       if (lib.libTopLevelExportElements.any(whereElement)) {
         return true;
       }
     }
 
-    return lib.exportedLibraries.any(_isExportedRecursivaly);
+    return lib.exportedLibraries.any((exported) {
+      final name = exported.lib.librarySource.fullName;
+      if (!visited.contains(name)) {
+        return _isExportedRecursivaly(exported, visited..add(name));
+      }
+      return false;
+    });
   }
 }
 
@@ -81,14 +93,14 @@ class LibraryElementWithVisibility {
   }
 
   Iterable<Element> get libTopLevelExportElements {
-    return lib.topLevelElements.where(isExported);
+    return lib.exportNamespace.definedNames.values.where(isExported);
   }
 
   /// based on exportedLibraries from analyzer-4.1.0
   /// lib/src/dart/element/element.dart
-  Iterable<LibraryElementWithVisibility> get exportedLibraries {
+  List<LibraryElementWithVisibility> get exportedLibraries {
     final libraries = <LibraryElementWithVisibility>{};
-    for (final export in lib.exports) {
+    for (final export in lib.libraryExports) {
       final library = export.exportedLibrary;
       if (library != null) {
         libraries.add(LibraryElementWithVisibility(